percent added

+++ manage.py
@@ -48,12 +48,22 @@ def delete():
     import reformed.dbconfig as config
     from sqlalchemy import MetaData
     print 'deleting database'
-    meta = MetaData()
-    meta.reflect(bind=config.engine)
 
-    for table in reversed(meta.sorted_tables):
-        print 'deleting %s...' % table.name
-        config.engine.execute(table.delete())
+    engine = config.engine.name
+
+    if engine == 'sqlite':
+        file = config.engine.url.database
+        if os.path.isfile(file):
+            if os.system("rm %s" % file):
+                print "ERROR: Couldn't delete database %s" % file
+                sys.exit(1)
+    else:
+        meta = MetaData()
+        meta.reflect(bind=config.engine)
+
+        for table in reversed(meta.sorted_tables):
+            print 'deleting %s...' % table.name
+            config.engine.execute(table.delete())
 
 def dump():
     print 'dumping data'
